back_end spring pour car rental project

+++ src/main/java/com/springboot/car_rental/dto/BookACarDto.java
@@ -12,8 +12,8 @@ public class BookACarDto {
     private Date fromDate;
     private Date toDate;
     private Long days;
-    private Long pricePerDay;
-    private Long price;
+    private double pricePerDay;
+    private double price;
     private BookCarStatus BookCarStatus;
     private Long carId;
     private Long userId;
